pyproject.toml
[tool.poetry]
name = "terrainman"
version = "0.0.38"
description = "A package for downloading and managing high-fidelity terrain files and solar irradiance data"
authors = ["Liam Robinson <robin502@purdue.edu>"]
readme = "README.md"
packages = [{include = "terrainman", from = "."}]
exclude = ['terrainman/__pycache']
documentation = "https://python-poetry.org/docs/"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: MacOS",
]

[tool.poetry.dependencies]
python = ">3.8"
numpy = ">=1.24.2"
scipy = ">=1.10.1"
python-dotenv = ">=1.0.0"
rasterio = ">=1.3.6"
netCDF4 = ">=1.6.3"
requests = ">=2.26.0"
bs4 = ">=0.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
